refactor(main): extract fallback campaign router wiring
